+++ src/main/java/org/betonquest/betonquest/api/bukkit/config/custom/handle/HandleModificationConfigurationSection.java
@@ -31,16 +31,24 @@ public class HandleModificationConfigurationSection extends ConfigurationSection
         this.handler = handler;
     }
 
+    @NotNull
+    @Override
+    public Map<String, Object> getValues(final boolean deep) {
+        final Map<String, Object> values = original.getValues(deep);
+        values.replaceAll((k, v) -> wrapModifiable(v));
+        return values;
+    }
+
     @Override
     public @Nullable
     Configuration getRoot() {
-        return new HandleModificationConfiguration(original.getRoot(), (ConfigurationModificationHandler) handler);
+        return (Configuration) wrapModifiable(original.getRoot());
     }
 
     @Override
     public @Nullable
     ConfigurationSection getParent() {
-        return original.getParent() == null ? null : (ConfigurationSection) wrapModifiable(original.getParent());
+        return (ConfigurationSection) wrapModifiable(original.getParent());
     }
 
     @Override
@@ -109,7 +117,13 @@ public class HandleModificationConfigurationSection extends ConfigurationSection
         handler.setInlineComments(original, path, comments);
     }
 
-    private Object wrapModifiable(final Object obj) {
+    /**
+     * Wraps a given object into an instance of this class.
+     *
+     * @param obj the raw object
+     * @return the wrapped object
+     */
+    protected Object wrapModifiable(final Object obj) {
         if (obj instanceof HandleModificationConfigurationSection) {
             return obj;
         }