+++ src/test/java/org/betonquest/betonquest/api/bukkit/config/custom/multi/MultiConfigurationSectionSplitWithConfigurationTest.java
@@ -8,7 +8,6 @@ import org.bukkit.configuration.file.YamlConfiguration;
 import java.io.File;
 import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -27,7 +26,7 @@ public class MultiConfigurationSectionSplitWithConfigurationTest extends MultiCo
         configs.put(YamlConfiguration.loadConfiguration(new File("src/test/resources/api/bukkit/multi/config3.yml")), "config1.yml");
         try {
             final MultiConfiguration multiConfiguration = new MultiConfiguration(new ArrayList<>(configs.keySet()));
-            multiConfiguration.setMultiDefaults(List.of(MultiConfigurationSplitTest.getDefault()));
+            multiConfiguration.setDefaults(MultiConfigurationSplitTest.getDefault());
             return multiConfiguration;
         } catch (final KeyConflictException e) {
             fail(e.resolvedMessage(configs), e);