+++ src/test/java/org/betonquest/betonquest/api/bukkit/config/custom/multi/MultiConfigurationTest.java
@@ -5,11 +5,9 @@ import org.bukkit.configuration.Configuration;
 import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.configuration.InvalidConfigurationException;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.function.Executable;
 
 import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -29,7 +27,7 @@ public class MultiConfigurationTest extends ConfigurationBaseTest {
             final MultiConfiguration multiConfiguration = new MultiConfiguration(new ArrayList<>(configs.keySet()));
             final Configuration defaults = defaultConfig.getDefaults();
             assertNotNull(defaults);
-            multiConfiguration.setMultiDefaults(List.of(defaults));
+            multiConfiguration.setDefaults(defaults);
             return multiConfiguration;
         } catch (final KeyConflictException e) {
             fail(e.resolvedMessage(configs), e);
@@ -37,59 +35,6 @@ public class MultiConfigurationTest extends ConfigurationBaseTest {
             fail(e);
         }
         return null;
-    }
-
-    private void assertThrowsUnmodifiableException(final Executable executable) {
-        final Exception exception = assertThrows(UnsupportedOperationException.class, executable);
-        assertEquals(MultiConfiguration.UNMODIFIABLE_MESSAGE, exception.getMessage());
-    }
-
-    @Test
-    @Override
-    public void testAddDefaultOnRootSection() {
-        assertThrowsUnmodifiableException(super::testAddDefaultOnRootSection);
-    }
-
-    @Test
-    @Override
-    public void testAddDefaultOnRootSectionOnExistingConfigPath() {
-        assertThrowsUnmodifiableException(super::testAddDefaultOnRootSectionOnExistingConfigPath);
-    }
-
-    @Test
-    @Override
-    public void testAddDefaultsAsConfiguration() {
-        assertThrowsUnmodifiableException(super::testAddDefaultsAsConfiguration);
-    }
-
-    @Test
-    @Override
-    public void testAddDefaultsAsConfigurationOnExistingConfigPath() {
-        assertThrowsUnmodifiableException(super::testAddDefaultsAsConfigurationOnExistingConfigPath);
-    }
-
-    @Test
-    @Override
-    public void testAddDefaultsAsMap() {
-        assertThrowsUnmodifiableException(super::testAddDefaultsAsMap);
-    }
-
-    @Test
-    @Override
-    public void testAddDefaultsAsMapOnExistingConfigPath() {
-        assertThrowsUnmodifiableException(super::testAddDefaultsAsMapOnExistingConfigPath);
-    }
-
-    @Test
-    @Override
-    public void testSetDefaults() {
-        assertThrowsUnmodifiableException(super::testSetDefaults);
-    }
-
-    @Test
-    @Override
-    public void testSetDefaultsOnExistingConfigPath() {
-        assertThrowsUnmodifiableException(super::testSetDefaultsOnExistingConfigPath);
     }
 
     @Test